% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inflate.R
\name{inflate_}
\alias{inflate_}
\title{Inflate a data frame}
\usage{
inflate_(data, var, domain = NULL)
}
\arguments{
\item{data(data.frame)}{A data frame to inflate}

\item{var(string)}{Column name specifying column to inflate over.
The column should be of class \code{Date}.}

\item{domain(Date)}{A vector of two dates (or two strings that can be
  coerced to dates). Used to specifiy the range to inflate over.
  The first date should specify the lower bound of the date range; the second
  date should specify the upper bound of the date range. If \code{NULL} the
  range is inferred by taking the minumum and maximum dates in the column.}
}
\description{
Inflate a data frame
}
\section{S3 generic}{

This function has methods for:
\itemize{
  \item \code{data.frame}
  \item \code{grouped_df}
}
}
\examples{
library(dplyr)

sales_data <- data_frame(
  date = as.Date(c("2014-01-01", "2014-01-03")),
  sales = c(10, 13)
)

inflate_(sales_data, "date")
inflate_(sales_data, "date", domain = c("2013-12-31", "2014-01-04"))

grp_sales_data <- data_frame(
  group = c("a", "a", "b", "b"),
  date = as.Date(c("2014-01-01", "2014-01-03", "2014-01-01", "2014-01-03")),
  sales = c(10, 13, 20, 30)
)

grp_sales_data \%>\%
  group_by(group) \%>\%
  inflate_("date")
}

